feat : 연차 조회

+++ src/main/java/com/example/workus/user/vo/User.java
@@ -31,6 +31,6 @@ public class User {
     private String positionName;
     private long deptNo;
     private String deptName;
-    private long roleNo;
+    private int roleNo;
     private String roleName;
 }
